Exclude pub build directory from analysis BUG=dartbug.com/15885 R=keertip@google.com
Review URL: https://codereview.chromium.org//126273002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@31568 260f80e4-7a28-3924-810f-c04153c831b5

+++ dart/editor/tools/plugins/com.google.dart.tools.core/src/com/google/dart/tools/core/internal/builder/AnalysisMarkerManager.java
@@ -314,10 +314,12 @@ public class AnalysisMarkerManager {
 
     //TODO(keertip): remove resource from queue
     try {
-      if (resource instanceof IContainer) {
-        resource.deleteMarkers(null, false, IResource.DEPTH_INFINITE);
-      } else {
-        resource.deleteMarkers(null, false, IResource.DEPTH_ZERO);
+      if (resource.isAccessible()) {
+        if (resource instanceof IContainer) {
+          resource.deleteMarkers(null, false, IResource.DEPTH_INFINITE);
+        } else {
+          resource.deleteMarkers(null, false, IResource.DEPTH_ZERO);
+        }
       }
     } catch (Exception e) {
       DartCore.logError(e);
